massive package restructure to make everything more obvious
