+++ bindings/go/descriptor/v2/descriptor.go
@@ -287,7 +287,7 @@ func (t *Timestamp) MarshalJSON() ([]byte, error) {
 	if y := t.Year(); y < 0 || y >= 10000 {
 		// RFC 3339 is clear that years are 4 digits exactly.
 		// See golang.org/issue/4556#c15 for more discussion.
-		return nil, errors.New("Time.Marshal: year outside of range [0,9999]")
+		return nil, errors.New("Time.MarshalJSON: year outside of range [0,9999]")
 	}
 
 	b := make([]byte, 0, len(time.RFC3339)+2)
