+++ bindings/go/plugin/manager/registries/plugins/decode_body.go
@@ -1,0 +1,16 @@
+package plugins
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+)
+
+func DecodeJSONRequestBody[T any](writer http.ResponseWriter, request *http.Request) (*T, error) {
+	pRequest := new(T)
+	if err := json.NewDecoder(request.Body).Decode(pRequest); err != nil {
+		writer.WriteHeader(http.StatusBadRequest)
+		return nil, fmt.Errorf("failed to decode request: %w", err)
+	}
+	return pRequest, nil
+}
