+++ bindings/go/plugin/manager/registries/plugins/errors.go
@@ -1,0 +1,16 @@
+package plugins
+
+import "net/http"
+
+type Error struct {
+	Err    error `json:"error"`
+	Status int   `json:"status"`
+}
+
+func NewError(err error, status int) *Error {
+	return &Error{Err: err, Status: status}
+}
+
+func (e *Error) Write(w http.ResponseWriter) {
+	http.Error(w, e.Err.Error(), e.Status)
+}
